Implemented piano quest
First version of piano quest. Instead of talking to people, all
interaction is through using items.
I ran into a problem where some things refered to the player object
created in player, and some to the
copy initialized in main. So instead I stopped making a new one in main
and had it set the name/description on the object we already created.